Move implementation out of __init__
